fix: HttpHandshakeInterceptor.java 파일 제거 및 chat 관련 설정 패키지 이동
