--- backend/src/main/java/com/NBE_4_5_2/Team5/global/config/HttpHandshakeInterceptor.kt
@@ -1,33 +1,0 @@
-package com.NBE_4_5_2.Team5.global.config
-
-import org.springframework.http.server.ServerHttpRequest
-import org.springframework.http.server.ServerHttpResponse
-import org.springframework.web.socket.WebSocketHandler
-import org.springframework.web.socket.server.HandshakeInterceptor
-
-class HttpHandshakeInterceptor : HandshakeInterceptor {
-
-    override fun beforeHandshake(
-        request: ServerHttpRequest,
-        response: ServerHttpResponse,
-        wsHandler: WebSocketHandler,
-        attributes: MutableMap<String, Any>
-    ): Boolean {
-        request.headers.getFirst("Cookie")
-            ?.split("; ")
-            ?.forEach { cookie ->
-                cookie.split("=")
-                    .takeIf { it.size == 2 }
-                    ?.let { pair -> attributes[pair[0]] = pair[1] }
-            }
-        return true
-    }
-
-    override fun afterHandshake(
-        request: ServerHttpRequest,
-        response: ServerHttpResponse,
-        wsHandler: WebSocketHandler,
-        exception: Exception?
-    ) {
-    }
-}
