+++ backend/src/main/java/com/NBE_4_5_2/Team5/global/config/chat/WebSocketConfig.kt
@@ -1,4 +1,4 @@
-package com.NBE_4_5_2.Team5.global.config
+package com.NBE_4_5_2.Team5.global.config.chat
 
 import com.NBE_4_5_2.Team5.global.handler.StompHandler
 import org.springframework.beans.factory.annotation.Value
